Update UserBoard and Menu
